requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
playwright==1.40.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
cryptography==41.0.7
python-keyring==24.2.0
openai==1.3.0
python-socketio==5.10.0
websockets==12.0
